+++ src/components/window/menubar/index.tsx
@@ -1,27 +1,24 @@
-import { Menubar, MenubarContent, MenubarGroup, MenubarItem, MenubarMenu, MenubarSeparator, MenubarShortcut, MenubarSub, MenubarSubContent, MenubarSubTrigger, MenubarTrigger } from "../../ui/menubar";
+import { Menubar, MenubarContent, MenubarItem, MenubarMenu, MenubarSub, MenubarSubContent, MenubarSubTrigger, MenubarTrigger } from "../../ui/menubar";
 import TranslatorActions from "@/actions/translator";
-import ReplaceTranslationPopup from "@/popups/modals/replace-translation";
-import BatchRenameKeysPopup from "@/popups/modals/batch-rename-keys";
 import AutoTranslatePopup from "@/popups/modals/auto-translate";
 import TransliterateScriptPopup from "@/popups/modals/transliterate-script";
 import { PROVIDER_NAMES } from "@/lib/constants";
 import { AutoTranslateProvider } from "@/schemas/types";
 import SpellCheckPopup from "@/popups/modals/spell-check"
-import CompareDifferencePopup from "@/popups/modals/compare-diff";
 import FileMenu from "./file-menu";
 import ViewMenu from "./view-menu";
-import EditActions from "@/actions/edit";
-import FindSubmenu from "./find-submenu";
 import { useAppTranslation } from "@/context/translation";
 import { toast } from "sonner";
+import EditMenu from "./edit-menu";
 
 export default function MenuBar(){
-     const {setTable, table, baseKeys} = useAppTranslation()
+     const {setTable, table, baseKeys, setIsDirty} = useAppTranslation()
      const removeUnusedKeys = () => {
           const res = TranslatorActions.removeUnusedKeys(table,baseKeys)
           if(res.success) {
                toast.success(res.success)
                setTable(res.data)
+               setIsDirty(true)
           }
      }
      const validateKeys = () => {
@@ -35,45 +32,7 @@ export default function MenuBar(){
      return (
           <Menubar className="h-full border-0 bg-transparent shadow-none rounded-none">
                <FileMenu/>
-               <MenubarMenu>
-                    <MenubarTrigger className="tracking-tight">Edit</MenubarTrigger>
-                    <MenubarContent>
-                         <MenubarGroup>
-                              <MenubarItem onClick={EditActions.undo}>Undo <MenubarShortcut>Ctrl+Z</MenubarShortcut></MenubarItem>
-                              <MenubarItem onClick={EditActions.redo}>Redo <MenubarShortcut>Ctrl+Y</MenubarShortcut></MenubarItem>
-                         </MenubarGroup>
-                         <MenubarSeparator />
-                         <MenubarGroup>
-                              <FindSubmenu/>
-                              <ReplaceTranslationPopup triggerButton={(
-                                   <MenubarItem onSelect={(e) => e.preventDefault()}>
-                                        Replace Translation
-                                        <MenubarShortcut>Ctrl+R</MenubarShortcut>
-                                   </MenubarItem>
-                              )}/>
-                              <BatchRenameKeysPopup triggerButton={(
-                                   <MenubarItem onSelect={(e) => e.preventDefault()}>
-                                        Batch Rename Keys
-                                        <MenubarShortcut>Ctrl+Shift+R</MenubarShortcut>
-                                   </MenubarItem>
-                              )}/>
-                         </MenubarGroup>
-                         <MenubarSeparator/>
-                         <MenubarGroup>
-                              <MenubarItem onClick={EditActions.selectUntranslated}>Select untranslated</MenubarItem>
-                              <CompareDifferencePopup triggerButton={(
-                                   <MenubarItem onSelect={(e) => e.preventDefault()}>Compare difference</MenubarItem>
-                              )}/>
-                         </MenubarGroup>
-                         <MenubarSeparator />
-                         <MenubarGroup>
-                              <MenubarItem onClick={EditActions.cut}>Cut <MenubarShortcut>Ctrl+X</MenubarShortcut></MenubarItem>
-                              <MenubarItem onClick={EditActions.copy}>Copy <MenubarShortcut>Ctrl+C</MenubarShortcut></MenubarItem>
-                              <MenubarItem onClick={EditActions.paste}>Paste <MenubarShortcut>Ctrl+V</MenubarShortcut></MenubarItem>
-                              <MenubarItem onClick={EditActions.selectAll}>Select All <MenubarShortcut>Ctrl+A</MenubarShortcut></MenubarItem>
-                         </MenubarGroup>
-                    </MenubarContent>
-               </MenubarMenu>
+               <EditMenu/>
                <ViewMenu/>
                <MenubarMenu>
                     <MenubarTrigger className="tracking-tight">Tools</MenubarTrigger>