+++ src/components/window/titlebar.tsx
@@ -9,6 +9,10 @@ import { ButtonGroup } from "@/components/ui/button-group"
 import MenuBar from "./menubar";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import { useState, useEffect } from "react";
+import { message } from "@tauri-apps/plugin-dialog";
+import { useAppTranslation } from "@/context/translation";
+import FileActions from "@/actions/file";
+import { toast } from "sonner";
 
 interface TitleBarProps{
      hideMenubar?: boolean,
@@ -18,14 +22,40 @@ interface TitleBarProps{
 export default function TitleBar({hideMaximize, hideMenubar, title}: TitleBarProps){
      const appWindow = getCurrentWindow()
      const [isMaximized, setIsMaximized] = useState(false)
-     // TODO: Use The Tauri dialog to save changes or cancel
-     const handleClose = async () => appWindow.close();
+     const {isDirty, files, setIsDirty, table} = useAppTranslation()
+     const handleClose = async () => {
+          if(!isDirty) {
+               await appWindow.close();
+               return
+          }
+          const paths = files.targetPath.split("\\")
+          const confirmation = await message(`Do you want to save changes to the translation "${paths[paths.length-1] || "Untitled"}"?`,{
+               title: "I18N Translator",
+               kind: "warning",
+               buttons: {
+                    yes: "Save",
+                    no: "Don't Save",
+                    cancel: "Cancel"
+               }
+          })
+          if (confirmation === "Save") {
+               const res = files.targetPath ? await FileActions.saveAll(table, files.targetPath) : await FileActions.saveAs(table)
+               if (res?.success) {
+                    setIsDirty(false)
+                    await appWindow.close()
+               }
+               if (res?.error) toast.error("Failed to save translation", {
+                    description: res.error,
+               })
+          }
+          if (confirmation === "Don't Save") await appWindow.close()
+     }
      const handleToggleMaximize = async () => {
           if(hideMaximize) return;
           await appWindow.toggleMaximize();
           setIsMaximized(await appWindow.isMaximized())
      }
-     const handleMinimize = async () => appWindow.minimize()
+     const handleMinimize = async () => await appWindow.minimize()
      useEffect(() => {
           const syncState = async () => {
                setIsMaximized(await appWindow.isMaximized())